don’t attach multiple listeners for the same path, but keep count
Basically the same one as here:
https://github.com/prescottprue/react-redux-firebase/issues/392 but this time one can turn it on or off.

The implementation that we currently have config.allowMultipleListeners might work for some, but in our case it was attaching way to many listeners to the same data. Despite the fact that the listeners returned the same data, we were sill dispatching set_listener and response events on the store and this caused a noticable lag.

if we turn off config.allowMultipleListeners, then everything seems fine to the moment where we attach the same listener on a scene and unmount the scene - the original listener also gets unmounted…

With this change we keep a count and don’t trigger the unsetListener function unless there are no more listeners for this path.