Break off composite query results from data
I think it's prudent to break off the composite query results from data.

This could be a breaking change for some people.  It's also nice to have at least something around as a backup if a listener is unset and then set again.

As such, I've made a semantic change and added another root collection.

`queries` contains the query definitions and their results.

`composite` contains the joining together of all of those queries (what was formerly stored in data).

`data` will be unmolested.