feat(cache-reducer): added new reducer for instant mutation changes
This effectively deprecated the need for both the dataReducer and the orderedReducer. The cache reducer also moves the populate functionality to run when the action is triggered, instead of in the selector. It also adds a 'fields' property on the query to reduce react over-rendering by ignoring unused data in the view.